allow keys to be entered in hex-encoded format
If a user needs to handle arbitrary keys, the existing mechanism of
directly using `fgets` is not sufficient.  It will terminate input on
newline, and is also unable to handle 0 (NUL) bytes in input.  To work
around this, this patch adds a `--hex` option to several operations
that instructs the password-reading code to expect a hex-encoded
string.  It will then decode the string into the original bytes before
use.

+++ src/sedcli_util.c
@@ -51,7 +51,7 @@ static void echo_enable()
     tcsetattr(1, 0, &term);
 }
 
-int get_password(char *pwd, uint8_t *len, int max)
+int get_password(char *pwd, uint16_t *len, uint16_t max)
 {
     size_t dest = max + 2;
     char temp[dest]; // note that this is VLA
